Deleted old liquid engine and implemented a barebones version without decision making
