--- src/nanopyx/liquid/__liquid_engine_base.py
@@ -1,345 +1,0 @@
-import os
-import timeit
-import yaml
-import datetime
-from functools import partial
-from pathlib import Path
-
-import numpy as np
-
-# This will in the future come from the Agent
-from .__njit__ import njit_works
-from .__opencl__ import opencl_works, devices
-
-__home_folder__ = os.path.expanduser("~")
-__benchmark_folder__ = os.path.join(__home_folder__, ".nanopyx")
-if not os.path.exists(__benchmark_folder__):
-    os.makedirs(__benchmark_folder__)
-
-class LiquidEngine_:
-
-    """
-    Base class for parts of the Nanopyx Liquid Engine
-    Vroom Vroom 
-    """
-
-    def __init__(self, testing:bool=False,
-                 opencl_:bool = False, unthreaded_:bool = False,
-                 threaded_:bool = False, threaded_static_:bool = False,
-                 threaded_dynamic_:bool = False, threaded_guided_:bool = False,
-                 python_:bool=False, njit_:bool=False) -> None:
-        """
-        Initialize the Liquid Engine
-        The Liquid Engine base class is inherited by children classes that implement specific methods
-
-        Engine responsabilities:
-        1. Store implemented run types;
-        2. Handle previous benchmarks and I/O;
-        2. When queried, benchmark all available run types;
-        3. Run a specific method using a selected run type;
-
-        Benchmark files have the following format:
-        The benchmark file is read as dict of dicts. 
-            BENCHMARK DICT FOR A SPECIFIC METHOD
-                |- RUN_TYPE #1
-                |      |- ARGS_REPR #1
-                |      |      |- [score, t2run#1, t2run#2, t2run#3, ...] last are newer. nan means fail
-                |      |- ARGS_REPR #2  
-                |      |      |- [score, t2run#1, t2run#2, t2run#3, ...] last are newer. nan means fail
-                |      (...)
-                |- RUN_TYPE #2 
-                (...)
-        """
-
-        # Start by checking available run types
-        self._run_types = {}
-        if opencl_ and opencl_works():
-            for d in devices:
-                self._run_types[f"OpenCL_{d['device'].name}"] = partial(self._run_opencl, device=d)
-        if threaded_:
-            self._run_types["Threaded"] = self._run_threaded
-        if unthreaded_:
-            self._run_types["Unthreaded"] = self._run_unthreaded
-        if threaded_static_:
-            self._run_types["Threaded_static"] = self._run_threaded_static
-        if threaded_dynamic_:
-            self._run_types["Threaded_dynamic"] = self._run_threaded_dynamic
-        if threaded_guided_:
-            self._run_types["Threaded_guided"] = self._run_threaded_guided
-        if python_:
-            self._run_types["Python"] = self._run_python
-        if njit_ and njit_works():
-            self._run_types["Numba"] = self._run_njit
-            # Try to trigger early compilation
-            try:
-                self._run_njit()
-            except TypeError:
-                print("Consider adding default arguments to the njit implementation to trigger early compilation")
-
-        self.testing = testing
-
-        # benchmarks file path
-        # e.g.: ~/.nanopyx/liquid/_le_interpolation_nearest_neighbor.cpython-310-darwin/ShiftAndMagnify.yml
-        base_path = os.path.join(__benchmark_folder__,"liquid",
-                                 os.path.split(os.path.splitext(inspect.getfile(self.__class__))[0])[1])
-        os.makedirs(base_path, exist_ok=True)
-        self._benchmark_filepath = os.path.join(base_path,self.__class__.__name__+".yml")
-
-        # Load config file if it exists, otherwise create an empty config
-        if not clear_benchmarks and os.path.exists(self._benchmark_filepath):
-            with open(self._benchmark_filepath) as f:
-                self._benchmarks = yaml.load(f, Loader=yaml.FullLoader)
-        else:
-            self._benchmarks = {}
-
-        # check if the cfg dictionary has a key for every available run type
-        for run_type_designation in self._run_types.keys():
-            if run_type_designation not in self._benchmarks:
-                self._benchmarks[run_type_designation] = {}
-        
-        # helper attribute for benchmarking function
-        self._last_time = None
-
-    def _run(self, *args, run_type:str, **kwargs):
-        """
-        Runs the function with the given args and kwargs
-
-        The code above does the following:
-        1. Check the specified run_type
-            - if str checks if the run type exists otherwise raise a NotImplementedError
-        2. It will run the _run_{run_type} function
-        3. It will return the result and the time taken to run
-
-        :param args: args for the function
-        :param run_type: the run type to use
-        :param kwargs: kwargs for the function
-        :return: the result and time taken
-        """
-
-        if run_type not in self._run_types:
-            print(f"Unexpected run type {run_type}")
-            raise NotImplementedError
-        
-        # try to run
-        try:
-            t_start = timeit.default_timer()
-            result = self._run_types[run_type](*args, **kwargs)
-            t2run = timeit.default_timer()-t_start
-        except Exception as e:
-            print(f"Unexpected error while trying to run {run_type}")
-            print(e)
-            print("Please try again with another run type")
-            result = None
-            t2run = np.inf
-
-        self._last_time = t2run
-        arg_repr, arg_score = self._get_args_repr_score(*args, **kwargs)
-        self._store_results(arg_repr, arg_score, run_type, t2run)
-
-        return result
-    
-    def benchmark(self,*args, **kwargs):
-        """
-        1. Run each available run type and record the run time and return value
-        2. Sort the run times from fastest to slowest
-        3. Compare each run type against each other, sorted by speed
-
-        :param args: args for the run method
-        :param kwargs: kwargs for the run method
-        :return:  a list of tuples containing the run time, run type name and optionally the return values
-        :rtype: [[run_time, run_type_name, return_value], ...]
-        """
-
-        # Create some lists to store runtimes and return values of run types
-        run_times = {}
-        returns = {}
-
-        # Run each run type and record the run time and return value
-        for run_type in self._run_types:
-
-            r = self._run(*args, run_type=run_type, **kwargs)
-
-            run_times[run_type] = self._last_time
-            
-            if self.testing: # Store return values if testing
-                returns[run_type] = r
-            else:
-                returns[run_type] = None
-        
-        # Sort run_times by value
-        speed_sort = []
-        for run_type in sorted(run_times, key=run_times.get, reverse=False):
-            speed_sort.append(
-                (
-                    run_times[run_type],
-                    run_type,
-                    returns[run_type],
-                )
-            )
-
-        print(f"Fastest run type: {speed_sort[0][1]}")
-        print(f"Slowest run type: {speed_sort[-1][1]}")
-
-        # Compare each run type against each other, sorted by speed
-        for i in range(len(speed_sort)):
-            if i not in run_times or run_times[i] is None:
-                continue
-            for j in range(i + 1, len(speed_sort)):
-                if j not in run_times or run_times[j] is None:
-                    continue
-
-                print(f"{speed_sort[i][1]} is {speed_sort[j][0]/speed_sort[i][0]:.2f} faster than {speed_sort[j][1]}")
-
-        return speed_sort
-
-    def _get_cl_code(self, file_name, cl_dp):
-        """
-        Retrieves the OpenCL code from the corresponding .cl file
-        """
-        cl_file = os.path.splitext(file_name)[0] + ".cl"
-        if not os.path.exists(cl_file):
-            cl_file = Path(__file__).parent / file_name
-
-        assert os.path.exists(cl_file), "Could not find OpenCL file: " + cl_file
-
-        kernel_str = open(cl_file).read()
-
-        if not cl_dp:
-            kernel_str = kernel_str.replace("double", "float")
-
-        return kernel_str
-
-    def _store_results(self, arg_repr, arg_score, run_type, t2run):
-        """
-        Stores the results of a run
-        """
-
-        # Check if the run type has been run, and if not create empty info
-        run_type_benchs = self._benchmarks[run_type]
-        if arg_repr not in run_type_benchs:
-            run_type_benchs[arg_repr] = [arg_score]
-
-        # Get the run info
-        c = run_type_benchs[arg_repr]
-
-        assert c[0] == arg_score, "arg_score mismatch"
-
-        # if run failed, t2run is np.inf
-        if np.isinf(t2run):
-            c.append(np.nan)
-        else:
-            c.append(t2run)
-
-        self._dump_run_times()
-
-    def _dump_run_times(self,):
-        """We might need to wrap this into a multiprocessing.Queue if we find it blocking"""
-        with open(self._benchmark_filepath, "w") as f:
-            yaml.dump(self._benchmarks, f)
-    
-    def _get_args_repr_score(self, *args, **kwargs):
-        """
-        Get a string representation of the args and kwargs and corresponding 'score' / 'norm'
-        The idea is that similar args have closer 'score'. Fuzzy logic
-
-        The code does the following:
-        1. It converts any args that are floats or ints to "number()" strings, and any args that are tensors to "shape()" strings
-        2. It converts any kwargs that are floats or ints to "number()" strings, and any kwargs that are tensors to "shape()" strings
-        3. The 'score' is given by the product of all the floats or ints and all the shape sizes. 
-
-        :return: the string representation of the args and kwargs
-        :rtype: str
-        """
-        _norm = 1
-        _args = []
-        for arg in args:
-            if type(arg) in (float, int):
-                _args.append(f"number({arg})")
-                _norm *= arg
-            elif hasattr(arg, "shape"):
-                _args.append(f"shape{arg.shape}")
-                _norm *= arg.size
-            else:
-                _args.append(arg)
-
-        _kwargs = {}
-        for k, v in kwargs.items():
-            if type(v) in (float, int):
-                _kwargs[k] = f"number({v})"
-                _norm *= arg
-            if hasattr(v, "shape"):
-                _kwargs[k] = f"shape{arg.shape}"
-                _norm *= arg.size
-            else:
-                _kwargs[k] = v
-
-        return repr((_args, _kwargs)), _norm
-
-    #####################################################
-    #                   RUN METHODS                     #
-    # THESE SHOULD ALWAYS BE OVERRIDEN BY CHILD CLASSES #
-    #####################################################
-
-    def run(self, *args, **kwargs):
-        """
-        Runs the function with the given args and kwargs
-        Should be overridden by the any class that inherits from this class
-        """
-        return self._run(*args, **kwargs)
-
-    def _run_opencl(*args, **kwargs):
-        """
-        Runs the OpenCL version of the function
-        Should be overridden by the any class that inherits from this class
-        """
-        pass
-
-    def _run_unthreaded(*args, **kwargs):
-        """
-        Runs the cython unthreaded version of the function
-        Should be overridden by the any class that inherits from this class
-        """
-        pass
-
-    def _run_threaded(*args, **kwargs):
-        """
-        Runs the cython threaded version of the function
-        Should be overridden by the any class that inherits from this class
-        """
-        pass
-
-    def _run_threaded_static(*args, **kwargs):
-        """
-        Runs the cython threaded static version of the function
-        Should be overridden by the any class that inherits from this class
-        """
-        pass
-
-    def _run_threaded_dynamic(*args, **kwargs):
-        """
-        Runs the cython threaded dynamic version of the function
-        Should be overridden by the any class that inherits from this class
-        """
-        pass
-
-    def _run_threaded_guided(*args, **kwargs):
-        """
-        Runs the cython threaded guided version of the function
-        Should be overridden by the any class that inherits from this class
-        """
-        pass
-
-    def _run_python(*args, **kwargs):
-        """
-        Runs the python version of the function
-        Should be overridden by the any class that inherits from this class
-        """
-        pass
-
-    def _run_njit(*args, **kwargs):
-        """
-        Runs the njit version of the function
-        Should be overridden by the any class that inherits from this class
-        """
-        pass
-
